Finalize changes to FTO Behavior
