+++ spec/fto-spec.js
@@ -1,5 +1,5 @@
 const Behaviors = require('../src/behaviors');
-const mockGithub = require('../src/utils').mockGithub;
+const mockGithub = require('./test-utils').mockGithub;
 
 const emptyResponse = 'You need to mention the name of a repository.';
 const existingResponse = 'publiclab/existing\n1 => My first fake issue\n2 => My second fake issue';
